change org name and move files, add manifest to jar via plugin
