+++ pom.xml
@@ -11,8 +11,8 @@
         <relativePath />
     </parent>
 
-    <groupId>org.example</groupId>
-    <artifactId>space-gambling</artifactId>
+    <groupId>org.space_gambling</groupId>
+    <artifactId>space_gambling</artifactId>
     <version>1.0-SNAPSHOT</version>
 
     <properties>
@@ -58,6 +58,13 @@
             <plugin>
                 <groupId>com.github.eirslett</groupId>
                 <artifactId>frontend-maven-plugin</artifactId>
+            </plugin>
+            <plugin>
+                <groupId>org.springframework.boot</groupId>
+                <artifactId>spring-boot-maven-plugin</artifactId>
+                <configuration>
+                    <mainClass>org.space_gambling.Runner</mainClass>
+                </configuration>                   
             </plugin>
         </plugins>
     </build>
