+++ src/main/java/org/space_gambling/Runner.java
@@ -1,4 +1,4 @@
-package org.example;
+package org.space_gambling;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
