+++ src/main/java/org/space_gambling/config/Constants.java
@@ -1,4 +1,4 @@
-package org.example.config;
+package org.space_gambling.config;
 
 import org.springframework.core.ParameterizedTypeReference;
