+++ src/main/java/org/space_gambling/config/SpaceTradersConfig.java
@@ -1,7 +1,7 @@
-package org.example.config;
+package org.space_gambling.config;
 
 import lombok.extern.slf4j.Slf4j;
-import org.example.service.TokenService;
+import org.space_gambling.service.TokenService;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
