+++ src/main/java/org/space_gambling/config/SpaceTradersProperties.java
@@ -1,4 +1,4 @@
-package org.example.config;
+package org.space_gambling.config;
 
 import lombok.Data;
 import org.springframework.boot.context.properties.ConfigurationProperties;
