+++ src/main/java/org/space_gambling/controller/HomeController.java
@@ -1,4 +1,4 @@
-package org.example.controller;
+package org.space_gambling.controller;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Controller;
