+++ src/main/java/org/space_gambling/controller/StartingController.java
@@ -1,10 +1,10 @@
-package org.example.controller;
+package org.space_gambling.controller;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.example.exception.SpaceTradersApiException;
-import org.example.service.AgentService;
-import org.example.service.FactionsService;
+import org.space_gambling.exception.SpaceTradersApiException;
+import org.space_gambling.service.AgentService;
+import org.space_gambling.service.FactionsService;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
