+++ src/main/java/org/space_gambling/service/AgentService.java
@@ -1,8 +1,8 @@
-package org.example.service;
+package org.space_gambling.service;
 
 import lombok.RequiredArgsConstructor;
-import org.example.exception.SpaceTradersApiException;
-import org.example.rest.SpaceTradersRestTemplate;
+import org.space_gambling.exception.SpaceTradersApiException;
+import org.space_gambling.rest.SpaceTradersRestTemplate;
 import org.springframework.http.HttpMethod;
 import org.springframework.stereotype.Service;
