+++ src/main/java/org/space_gambling/service/TokenService.java
@@ -1,4 +1,4 @@
-package org.example.service;
+package org.space_gambling.service;
 
 import io.github.cdimascio.dotenv.Dotenv;
 import lombok.extern.slf4j.Slf4j;
